fix(frontend): 修复 PR 前端 lint 失败
