+++ Frontend/radish.console/src/pages/Moderation/moderationPageHelpers.ts
@@ -1,4 +1,3 @@
-import { Tag } from '@radish/ui';
 import {
   type ContentReportQueueItemVo,
   type UserModerationActionVo,
@@ -150,7 +149,7 @@ export interface ModerationTargetNavigationStateInput extends ModerationTargetNa
   navigationMessage?: string | null;
 }
 
-interface ModerationTargetDisplayInput extends ModerationTargetNavigationStateInput {
+export interface ModerationTargetDisplayInput extends ModerationTargetNavigationStateInput {
   snapshotTitle?: string | null;
   snapshotSummary?: string | null;
   snapshotIsPersisted?: boolean;
@@ -163,7 +162,7 @@ export function getReasonTypeLabel(value: string | null | undefined): string {
   return value ? (REASON_TYPE_LABEL_MAP[value] ?? value) : '未分类';
 }
 
-function getTargetTypeLabel(value: string | null | undefined): string {
+export function getTargetTypeLabel(value: string | null | undefined): string {
   return value ? (TARGET_TYPE_LABEL_MAP[value] ?? value) : '未知目标';
 }
 
@@ -175,32 +174,6 @@ export function toPositiveLongString(value: string): string | undefined {
   const trimmed = value.trim();
   return /^[1-9]\d*$/.test(trimmed) ? trimmed : undefined;
 }
-
-export const renderReportStatus = (value: string) => {
-  switch (value) {
-    case 'Approved':
-      return <Tag color="success">已通过</Tag>;
-    case 'Rejected':
-      return <Tag color="error">已驳回</Tag>;
-    default:
-      return <Tag color="processing">待审核</Tag>;
-  }
-};
-
-export const renderActionType = (value: string) => {
-  switch (value) {
-    case 'Mute':
-      return <Tag color="orange">禁言</Tag>;
-    case 'Ban':
-      return <Tag color="red">封禁</Tag>;
-    case 'Unmute':
-      return <Tag color="green">解除禁言</Tag>;
-    case 'Unban':
-      return <Tag color="cyan">解除封禁</Tag>;
-    default:
-      return <Tag>无动作</Tag>;
-  }
-};
 
 export function getActionTypeText(value: string | null | undefined): string {
   switch (value) {
@@ -265,7 +238,7 @@ export function canOpenChatTarget(
     && messageId > 0;
 }
 
-function resolveNavigationStatusLabel(status: string | null | undefined): { color: string; label: string } {
+export function resolveNavigationStatusLabel(status: string | null | undefined): { color: string; label: string } {
   switch (status) {
     case 'Fallback':
       return { color: 'warning', label: '已降级' };
@@ -276,66 +249,6 @@ function resolveNavigationStatusLabel(status: string | null | undefined): { colo
     default:
       return { color: 'success', label: '可回看' };
   }
-}
-
-function renderTargetNavigationState(status: string | null | undefined, messageText: string | null | undefined) {
-  const statusMeta = resolveNavigationStatusLabel(status);
-
-  return (
-    <div className="moderation-target__section">
-      <div className="moderation-target__section-label">当前状态</div>
-      <div className="moderation-target__state">
-        <Tag color={statusMeta.color}>{statusMeta.label}</Tag>
-        {messageText ? <div className="moderation-target__state-message">{messageText}</div> : null}
-      </div>
-    </div>
-  );
-}
-
-function renderSnapshotSection(input: ModerationTargetDisplayInput) {
-  const hasSnapshotTitle = !!input.snapshotTitle?.trim();
-  const hasSnapshotSummary = !!input.snapshotSummary?.trim();
-
-  if (!hasSnapshotTitle && !hasSnapshotSummary && !input.snapshotIsPersisted) {
-    return null;
-  }
-
-  return (
-    <div className="moderation-target__section">
-      <div className="moderation-target__section-label">
-        {input.snapshotIsPersisted ? '创建时快照' : '目标摘要（旧数据兼容）'}
-      </div>
-      {hasSnapshotTitle ? <div className="moderation-target__snapshot-title">{input.snapshotTitle}</div> : null}
-      {hasSnapshotSummary ? <div className="moderation-target__snapshot-summary">{input.snapshotSummary}</div> : null}
-      {!hasSnapshotTitle && !hasSnapshotSummary ? <div className="moderation-target__empty">未保留文本摘要</div> : null}
-    </div>
-  );
-}
-
-export function renderModerationTarget(input: ModerationTargetDisplayInput) {
-  return (
-    <div className="moderation-target">
-      <div className="moderation-target__identity">{getTargetTypeLabel(input.targetType)} #{input.targetContentId ?? '-'}</div>
-      {input.targetType === 'Comment' && input.targetPostId ? (
-        <div className="moderation-target__meta">
-          帖子 #{input.targetPostId} · 评论 #{input.targetCommentId ?? input.targetContentId}
-        </div>
-      ) : null}
-      {input.targetType === 'PostQuickReply' && input.targetPostId ? (
-        <div className="moderation-target__meta">所属帖子 #{input.targetPostId}</div>
-      ) : null}
-      {input.targetType === 'ChatMessage' && input.targetChannelId ? (
-        <div className="moderation-target__meta">
-          频道 #{input.targetChannelId} · 消息 #{input.targetMessageId ?? input.targetContentId}
-        </div>
-      ) : null}
-      {renderSnapshotSection(input)}
-      {renderTargetNavigationState(input.navigationStatus, input.navigationMessage)}
-      {input.showTargetUser ? (
-        <div className="moderation-target__user">{input.targetUserName || `用户 ${input.targetUserId}`}</div>
-      ) : null}
-    </div>
-  );
 }
 
 export function buildQueueTargetDisplayInput(record: ContentReportQueueItemVo): ModerationTargetDisplayInput {
